Consolidate to single E2E test with per-test output directories
- Remove 6 scaffolding/smoke tests (VsCodeLaunchesAndRendersWorkbench,
  Hex1bTerminalStartsWithWebSocketPort, RemoteTerminalSessionCanSendAndReceiveText,
  InstallAspireCliViaTerminal, AspireNewCreatesProjectInteractively)
- Keep only ExtensionShowsResourcesFromRunningAppHost (full integration test)
- Rename class SmokeTests → ExtensionEndToEndTests
- Refactor fixture: each test gets its own output directory under
  artifacts/testresults/extension-e2e/{TestName}/ with all artifacts
  (screenshots, traces, recordings, videos) in one place
- Number screenshot filenames (01-cli-installed, 02-project-created, etc.)
  for clear chronological ordering
- Each test creates its own browser context (videos/traces scoped per-test)
