+++ tests/Aspire.Extension.EndToEndTests/ExtensionEndToEndTests.cs
@@ -1,0 +1,564 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Extension.EndToEndTests.Infrastructure;
+using Aspire.Extension.EndToEndTests.Infrastructure.Hex1b;
+using Microsoft.Playwright;
+using Xunit;
+
+namespace Aspire.Extension.EndToEndTests;
+
+/// <summary>
+/// End-to-end tests for the Aspire VS Code extension running in a browser
+/// via Docker + Playwright + Hex1b terminal automation.
+/// </summary>
+public sealed class ExtensionEndToEndTests : IClassFixture<VsCodeWebFixture>, IAsyncDisposable
+{
+    private readonly VsCodeWebFixture _fixture;
+    private readonly ITestOutputHelper _output;
+
+    public ExtensionEndToEndTests(VsCodeWebFixture fixture, ITestOutputHelper output)
+    {
+        _fixture = fixture;
+        _output = output;
+    }
+
+    [Fact]
+    public async Task ExtensionShowsResourcesFromRunningAppHost()
+    {
+        var testDir = _fixture.CreateTestOutputDir(nameof(ExtensionShowsResourcesFromRunningAppHost));
+
+        // Skip if no locally-built artifacts are available
+        var artifacts = _fixture.Artifacts;
+        if (artifacts is null)
+        {
+            _output.WriteLine("SKIP: No locally-built artifacts detected. " +
+                AspireBuildArtifacts.DescribeMissing(FindRepoRoot()));
+            return;
+        }
+
+        _output.WriteLine("=== Full Integration Test: Extension + AppHost Resources ===");
+        _output.WriteLine($"  CLI:      {artifacts.CliPublishDirectory}");
+        _output.WriteLine($"  VSIX:     {artifacts.VsixPath}");
+        _output.WriteLine($"  Packages: {artifacts.PackagesDirectory}");
+
+        var page = await _fixture.CreatePageAsync(testDir);
+
+        try
+        {
+            // Wait for VS Code and open terminal
+            await page.WaitForSelectorAsync(".monaco-workbench", new() { Timeout = 120_000 });
+            await page.Keyboard.PressAsync("Escape");
+            await Task.Delay(500);
+            await page.Keyboard.PressAsync("Escape");
+            await Task.Delay(1000);
+
+            _output.WriteLine("Opening integrated terminal...");
+            await page.Keyboard.PressAsync("Control+Backquote");
+            await page.WaitForSelectorAsync(".terminal-wrapper", new() { Timeout = 60_000 });
+            await Task.Delay(2000);
+
+            // Clean up previous state
+            await _fixture.Container.ExecAsync("rm -rf $HOME/.aspire /tmp/integration-* /root/MyAspireApp");
+            await _fixture.Container.ExecAsync("pkill -f 'hex1b terminal' 2>/dev/null");
+            _fixture.Container.ResetHex1bPorts();
+            await Task.Delay(1000);
+
+            // Maximize the terminal panel
+            await MaximizeTerminalPanelAsync(page);
+
+            // Start hex1b with passthru mode and asciinema recording
+            var recordPath = "/tmp/integration-test.cast";
+            var (containerPort, wsUri) = _fixture.Container.AllocateHex1bPort();
+            var hex1bCmd = $"hex1b terminal start --passthru --port {containerPort} --bind 0.0.0.0 --record {recordPath} -- /bin/bash";
+            _output.WriteLine($"Starting hex1b: {hex1bCmd}");
+            await page.Keyboard.TypeAsync(hex1bCmd, new() { Delay = 20 });
+            await page.Keyboard.PressAsync("Enter");
+
+            // Connect remote terminal session
+            await _fixture.Container.WaitForHex1bAsync(wsUri, timeout: TimeSpan.FromSeconds(60));
+            await using var session = await RemoteTerminalSession.ConnectAsync(wsUri);
+            _output.WriteLine("Remote terminal session connected");
+
+            // Wait for shell prompt
+            await session.WaitForTextAsync("~#", timeout: TimeSpan.FromSeconds(10));
+
+            // Verify the session works
+            var marker = $"READY_{Guid.NewGuid():N}";
+            await session.SendTextAsync($"echo {marker}\r");
+            var echoOk = await session.WaitForTextAsync(marker, timeout: TimeSpan.FromSeconds(10));
+            Assert.True(echoOk, "Remote terminal session echo test failed");
+            await session.WaitForTextAsync("~#", timeout: TimeSpan.FromSeconds(5));
+
+            // ===== Phase 1: Install local CLI =====
+            _output.WriteLine("--- Phase 1: Installing local CLI ---");
+            await session.SendTextAsync("mkdir -p ~/.aspire/bin && cp /opt/aspire-cli/aspire ~/.aspire/bin/aspire && chmod +x ~/.aspire/bin/aspire\r");
+            await session.WaitForTextAsync("~#", timeout: TimeSpan.FromSeconds(10));
+
+            // Set PATH and environment variables
+            await session.SendTextAsync("export PATH=~/.aspire/bin:$PATH ASPIRE_PLAYGROUND=true BUILT_NUGETS_PATH=/opt/aspire/packages TERM=xterm DOTNET_CLI_TELEMETRY_OPTOUT=true DOTNET_SKIP_FIRST_TIME_EXPERIENCE=true DOTNET_GENERATE_ASPNET_CERTIFICATE=false\r");
+            await session.WaitForTextAsync("~#", timeout: TimeSpan.FromSeconds(5));
+
+            // Verify CLI version
+            await session.SendTextAsync("aspire --version\r");
+            await Task.Delay(3000);
+
+            var versionResult = await _fixture.Container.ExecAsync(
+                "export PATH=$HOME/.aspire/bin:$PATH && aspire --version",
+                timeout: TimeSpan.FromSeconds(30));
+            var version = versionResult.StdOut.Trim();
+            _output.WriteLine($"aspire --version: {version}");
+            Assert.Equal(0, versionResult.ExitCode);
+            Assert.NotEmpty(version);
+
+            await ScreenshotAsync(page, testDir, "01-cli-installed.png");
+            _output.WriteLine($"✓ CLI installed: {version}");
+
+            // ===== Phase 2: Create project with aspire new =====
+            _output.WriteLine("--- Phase 2: Creating project with aspire new ---");
+            await session.SendTextAsync("aspire new\r");
+
+            // Wait for template selection prompt
+            _output.WriteLine("Waiting for template selection prompt...");
+            var templatePrompt = await session.WaitForTextAsync("Select a template", timeout: TimeSpan.FromSeconds(60));
+            if (!templatePrompt)
+            {
+                DumpScreen(session, "Template prompt NOT found");
+            }
+            Assert.True(templatePrompt, "Expected 'Select a template' prompt");
+            _output.WriteLine("Template selection prompt appeared");
+
+            // Select the first template (Starter App) by pressing Enter
+            await session.SendEnterAsync();
+
+            // Wait for project name prompt
+            var namePrompt = await session.WaitForTextAsync("project name", timeout: TimeSpan.FromSeconds(15));
+            Assert.True(namePrompt, "Expected project name prompt");
+
+            // Type project name and press Enter
+            var projectName = "MyAspireApp";
+            await session.SendTextAsync(projectName);
+            await Task.Delay(500);
+            await session.SendEnterAsync();
+
+            // Wait for output path prompt
+            var pathPrompt = await session.WaitForTextAsync("output path", timeout: TimeSpan.FromSeconds(15));
+            Assert.True(pathPrompt, "Expected output path prompt");
+
+            // Accept the default output path
+            await session.SendEnterAsync();
+
+            // Handle additional template-specific prompts and wait for success
+            _output.WriteLine("Answering prompts and waiting for project creation...");
+            string? created = null;
+            var createDeadline = DateTime.UtcNow + TimeSpan.FromMinutes(5);
+            var promptsAnswered = 0;
+
+            while (DateTime.UtcNow < createDeadline)
+            {
+                await Task.Delay(2000);
+
+                var fullText = session.GetFullText();
+
+                if (fullText.Contains("created successfully", StringComparison.OrdinalIgnoreCase) ||
+                    fullText.Contains("project at", StringComparison.Ordinal))
+                {
+                    created = "Created";
+                    break;
+                }
+
+                var screenText = session.GetScreenText();
+                if (screenText.Contains("Type to search", StringComparison.Ordinal))
+                {
+                    promptsAnswered++;
+                    _output.WriteLine($"  Additional prompt #{promptsAnswered} — pressing Enter");
+                    await session.SendEnterAsync();
+                    continue;
+                }
+            }
+
+            Assert.NotNull(created);
+            _output.WriteLine("✓ Project created successfully");
+            await ScreenshotAsync(page, testDir, "02-project-created.png");
+
+            // Handle the "Would you like to configure AI agent environments?" prompt
+            _output.WriteLine("Checking for AI agent environments prompt...");
+            var aiPromptDeadline = DateTime.UtcNow + TimeSpan.FromSeconds(30);
+            while (DateTime.UtcNow < aiPromptDeadline)
+            {
+                await Task.Delay(1000);
+                var screenNow = session.GetScreenText();
+
+                if (screenNow.Contains("[y/n]", StringComparison.Ordinal) ||
+                    screenNow.Contains("AI agent", StringComparison.OrdinalIgnoreCase) ||
+                    screenNow.Contains("configure", StringComparison.OrdinalIgnoreCase))
+                {
+                    _output.WriteLine("AI agent environments prompt detected — declining with 'n'");
+                    await session.SendTextAsync("n\r");
+                    break;
+                }
+
+                if (screenNow.Contains("~#", StringComparison.Ordinal) ||
+                    screenNow.Contains("root@", StringComparison.Ordinal))
+                {
+                    _output.WriteLine("Already at shell prompt — no AI agent prompt");
+                    break;
+                }
+            }
+
+            await session.WaitForAnyTextAsync(["~#", "root@"], timeout: TimeSpan.FromSeconds(15));
+            _output.WriteLine("Shell prompt ready");
+
+            // ===== Phase 2b: Patch SDK version to use dev packages =====
+            _output.WriteLine("--- Phase 2b: Patching SDK version to use dev packages ---");
+            var patchResult = await _fixture.Container.ExecAsync(
+                $"find /root/{projectName} -name '*.csproj' -exec sed -i 's|Aspire.AppHost.Sdk/[0-9.]*\"|Aspire.AppHost.Sdk/{version}\"|g' {{}} \\; && " +
+                $"head -1 /root/{projectName}/{projectName}.AppHost/{projectName}.AppHost.csproj",
+                timeout: TimeSpan.FromSeconds(10));
+            _output.WriteLine($"Patched AppHost SDK → {patchResult.StdOut.Trim()}");
+
+            // ===== Phase 3: Open project folder in VS Code =====
+            _output.WriteLine("--- Phase 3: Opening project folder in VS Code ---");
+
+            var folderUrl = $"{_fixture.Url}/?folder=/root/{projectName}";
+            _output.WriteLine($"Navigating to: {folderUrl}");
+            await page.GotoAsync(folderUrl);
+
+            // VS Code reloads when opening a folder — wait for workbench
+            _output.WriteLine("Waiting for VS Code to reload with project folder...");
+            await page.WaitForSelectorAsync(".monaco-workbench", new() { Timeout = 120_000 });
+            await Task.Delay(5000);
+
+            // Dismiss any trust dialog that may appear
+            var trustButton = await page.QuerySelectorAsync("text=Yes, I trust the authors");
+            if (trustButton is not null)
+            {
+                await trustButton.ClickAsync();
+                _output.WriteLine("Dismissed workspace trust dialog");
+                await Task.Delay(2000);
+            }
+
+            await ScreenshotAsync(page, testDir, "03-folder-opened.png");
+
+            // The terminal closed during reload — reopen it
+            _output.WriteLine("Reopening terminal after folder change...");
+            await page.Keyboard.PressAsync("Escape");
+            await Task.Delay(500);
+            await page.Keyboard.PressAsync("Control+Backquote");
+            await page.WaitForSelectorAsync(".terminal-wrapper", new() { Timeout = 60_000 });
+            await Task.Delay(2000);
+
+            // Clean up old hex1b and start fresh
+            await _fixture.Container.ExecAsync("pkill -f 'hex1b terminal' 2>/dev/null");
+            _fixture.Container.ResetHex1bPorts();
+            await Task.Delay(1000);
+
+            // Maximize the terminal panel
+            await MaximizeTerminalPanelAsync(page);
+
+            // Start hex1b again (since terminal was recreated)
+            var recordPath2 = "/tmp/integration-test-phase2.cast";
+            var (containerPort2, wsUri2) = _fixture.Container.AllocateHex1bPort();
+            var hex1bCmd2 = $"hex1b terminal start --passthru --port {containerPort2} --bind 0.0.0.0 --record {recordPath2} -- /bin/bash";
+            _output.WriteLine($"Starting hex1b: {hex1bCmd2}");
+            await page.Keyboard.TypeAsync(hex1bCmd2, new() { Delay = 20 });
+            await page.Keyboard.PressAsync("Enter");
+
+            // Connect new remote terminal session
+            await _fixture.Container.WaitForHex1bAsync(wsUri2, timeout: TimeSpan.FromSeconds(60));
+            await using var session2 = await RemoteTerminalSession.ConnectAsync(wsUri2);
+            _output.WriteLine("Remote terminal session reconnected");
+
+            await session2.WaitForAnyTextAsync(["~#", "root@", "MyAspireApp#"], timeout: TimeSpan.FromSeconds(10));
+
+            // Re-set environment variables in the new shell
+            await session2.SendTextAsync("export PATH=~/.aspire/bin:$PATH ASPIRE_PLAYGROUND=true BUILT_NUGETS_PATH=/opt/aspire/packages TERM=xterm DOTNET_CLI_TELEMETRY_OPTOUT=true DOTNET_SKIP_FIRST_TIME_EXPERIENCE=true DOTNET_GENERATE_ASPNET_CERTIFICATE=false\r");
+            await session2.WaitForAnyTextAsync(["~#", "root@", "MyAspireApp#"], timeout: TimeSpan.FromSeconds(5));
+
+            _output.WriteLine("✓ Project folder opened in VS Code");
+
+            // ===== Phase 4: Set up NuGet config + restore =====
+            _output.WriteLine("--- Phase 4: Setting up NuGet config ---");
+
+            await session2.SendTextAsync($"cp /opt/aspire/nuget.config /root/{projectName}/nuget.config\r");
+            await session2.WaitForAnyTextAsync(["~#", "root@"], timeout: TimeSpan.FromSeconds(5));
+
+            _output.WriteLine("Running dotnet restore...");
+            await session2.SendTextAsync($"cd /root/{projectName} && dotnet restore\r");
+
+            var restoreDone = await session2.WaitForAnyTextAsync(
+                ["Restore complete", "Build succeeded", "Nothing to do", projectName + "#"],
+                timeout: TimeSpan.FromMinutes(5),
+                includeScrollback: true);
+            _output.WriteLine($"Restore signal: {restoreDone ?? "TIMEOUT"}");
+
+            await session2.SendTextAsync("echo RESTORE_EXIT=$?\r");
+            var restoreExitOk = await session2.WaitForTextAsync("RESTORE_EXIT=0", timeout: TimeSpan.FromSeconds(10));
+            if (!restoreExitOk)
+            {
+                DumpScreen(session2, "Restore may have failed");
+                _output.WriteLine("WARNING: dotnet restore may have failed — continuing anyway");
+            }
+
+            await ScreenshotAsync(page, testDir, "04-restored.png");
+            _output.WriteLine("✓ NuGet restore complete");
+
+            // ===== Phase 5: Run aspire start =====
+            _output.WriteLine("--- Phase 5: Starting AppHost with aspire start ---");
+            await session2.SendTextAsync($"cd /root/{projectName} && aspire start\r");
+
+            var startedOk = await session2.WaitForAnyTextAsync(
+                ["AppHost started successfully", "Dashboard"],
+                timeout: TimeSpan.FromMinutes(5),
+                includeScrollback: true);
+
+            if (startedOk is null)
+            {
+                DumpScreen(session2, "aspire start - did not succeed");
+                Assert.Fail("aspire start did not report success within 5 minutes");
+            }
+            _output.WriteLine($"✓ AppHost started — signal: {startedOk}");
+
+            await ScreenshotAsync(page, testDir, "05-aspire-started.png");
+
+            // ===== Phase 5b: Wait for resources to be discoverable =====
+            _output.WriteLine("--- Phase 5b: Waiting for resources via backchannel ---");
+
+            var backchannelCheck = await _fixture.Container.ExecAsync(
+                "ls -la /root/.aspire/cli/backchannels/ 2>/dev/null",
+                timeout: TimeSpan.FromSeconds(5));
+            _output.WriteLine($"Backchannel sockets:\n{backchannelCheck.StdOut.Trim()}");
+
+            var resourcesDiscovered = false;
+            for (var attempt = 0; attempt < 12; attempt++)
+            {
+                await Task.Delay(5000);
+                var psResult = await _fixture.Container.ExecAsync(
+                    "export PATH=$HOME/.aspire/bin:$PATH && aspire ps --format json --resources 2>/dev/null",
+                    timeout: TimeSpan.FromSeconds(15));
+                var psOutput = psResult.StdOut.Trim();
+
+                if (psOutput.Contains("\"name\":", StringComparison.Ordinal))
+                {
+                    resourcesDiscovered = true;
+                    _output.WriteLine($"✓ Resources discovered (attempt {attempt + 1})");
+
+                    foreach (var line in psOutput.Split('\n'))
+                    {
+                        if (line.Contains("\"displayName\":", StringComparison.Ordinal))
+                        {
+                            _output.WriteLine($"  Resource: {line.Trim()}");
+                        }
+                    }
+                    break;
+                }
+                _output.WriteLine($"  Attempt {attempt + 1}: resources not yet available");
+            }
+
+            if (!resourcesDiscovered)
+            {
+                var procResult = await _fixture.Container.ExecAsync(
+                    "ps aux | grep -i apphost | head -5",
+                    timeout: TimeSpan.FromSeconds(5));
+                _output.WriteLine($"AppHost processes:\n{procResult.StdOut}");
+
+                var logResult = await _fixture.Container.ExecAsync(
+                    "cat $(ls -t /root/.aspire/logs/cli_*detach* 2>/dev/null | head -1) 2>/dev/null | tail -30",
+                    timeout: TimeSpan.FromSeconds(10));
+                _output.WriteLine($"AppHost detach log:\n{logResult.StdOut}");
+
+                Assert.Fail("Resources not discoverable via aspire ps --resources after 60 seconds");
+            }
+
+            // ===== Phase 6: Verify extension shows resources =====
+            _output.WriteLine("--- Phase 6: Verifying extension tree view ---");
+
+            // Un-maximize the terminal so we can see the activity bar and sidebar
+            await MaximizeTerminalPanelAsync(page);
+            await Task.Delay(1000);
+
+            // Click the Aspire icon in the activity bar to open the panel
+            _output.WriteLine("Looking for Aspire activity bar icon...");
+            var aspireIcon = await page.QuerySelectorAsync("[aria-label*='Aspire' i]")
+                          ?? await page.QuerySelectorAsync("[id*='aspire' i]")
+                          ?? await page.QuerySelectorAsync("[id*='aspire-panel']");
+
+            if (aspireIcon is not null)
+            {
+                await aspireIcon.ClickAsync();
+                _output.WriteLine("Clicked Aspire activity bar icon");
+                await Task.Delay(2000);
+            }
+            else
+            {
+                _output.WriteLine("Aspire icon not found — using command palette");
+                await page.Keyboard.PressAsync("Control+Shift+KeyP");
+                await Task.Delay(500);
+                await page.Keyboard.TypeAsync("Aspire: Focus on Running AppHosts View", new() { Delay = 30 });
+                await Task.Delay(500);
+                await page.Keyboard.PressAsync("Enter");
+                await Task.Delay(2000);
+            }
+
+            await ScreenshotAsync(page, testDir, "06-aspire-panel.png");
+
+            // Click Refresh to trigger immediate re-discovery via describe --follow
+            _output.WriteLine("Clicking Refresh to trigger re-discovery...");
+            var refreshButton = await page.QuerySelectorAsync("[aria-label='Refresh']")
+                             ?? await page.QuerySelectorAsync("a[title='Refresh']");
+            if (refreshButton is not null)
+            {
+                await refreshButton.ClickAsync();
+                _output.WriteLine("Clicked Refresh button");
+            }
+
+            // Retry loop: wait for tree items to appear (resources from backchannel RPC)
+            _output.WriteLine("Waiting for tree items to appear...");
+            IReadOnlyList<Microsoft.Playwright.IElementHandle> treeItems = [];
+            for (var attempt = 0; attempt < 12; attempt++)
+            {
+                await Task.Delay(5000);
+                treeItems = await page.QuerySelectorAllAsync("[role='treeitem']");
+                _output.WriteLine($"  Attempt {attempt + 1}: {treeItems.Count} tree items");
+                if (treeItems.Count > 0)
+                {
+                    break;
+                }
+
+                if (attempt % 3 == 2)
+                {
+                    refreshButton = await page.QuerySelectorAsync("[aria-label='Refresh']")
+                                 ?? await page.QuerySelectorAsync("a[title='Refresh']");
+                    if (refreshButton is not null)
+                    {
+                        await refreshButton.ClickAsync();
+                        _output.WriteLine("  Re-clicked Refresh");
+                    }
+                }
+            }
+
+            await ScreenshotAsync(page, testDir, "07-resources.png");
+
+            // Log all tree item labels
+            var allLabels = new List<string>();
+            foreach (var item in treeItems)
+            {
+                var label = await item.GetAttributeAsync("aria-label") ?? await item.InnerTextAsync();
+                allLabels.Add(label);
+                _output.WriteLine($"  Tree item: '{label}'");
+            }
+            _output.WriteLine($"All tree item labels: [{string.Join(", ", allLabels)}]");
+
+            // Assert resources are visible
+            if (treeItems.Count == 0)
+            {
+                await ScreenshotAsync(page, testDir, "07-no-resources-debug.png");
+
+                // Dump extension Output channel for debugging
+                await page.Keyboard.PressAsync("Control+Shift+KeyP");
+                await Task.Delay(500);
+                await page.Keyboard.TypeAsync("Output: Show Output Channel...", new() { Delay = 30 });
+                await Task.Delay(500);
+                await page.Keyboard.PressAsync("Enter");
+                await Task.Delay(500);
+                await page.Keyboard.TypeAsync("Aspire", new() { Delay = 30 });
+                await Task.Delay(500);
+                await page.Keyboard.PressAsync("Enter");
+                await Task.Delay(2000);
+                await ScreenshotAsync(page, testDir, "07-aspire-output-channel.png");
+
+                Assert.Fail("No tree items found in the extension panel after 60s of retries");
+            }
+
+            _output.WriteLine($"✓ Extension shows {treeItems.Count} resource(s) in the tree view");
+
+            // Linger for 5 seconds to capture final state in video
+            await Task.Delay(5000);
+
+            await ScreenshotAsync(page, testDir, "08-final.png");
+            _output.WriteLine("=== Full Integration Test Complete ===");
+
+            // Copy asciinema recordings from both sessions
+            var castPath1 = Path.Combine(testDir, "integration-test-phase1.cast");
+            await _fixture.Container.CopyFromContainerAsync(recordPath, castPath1);
+            if (File.Exists(castPath1))
+            {
+                _output.WriteLine($"Asciinema recording (phase 1): {castPath1}");
+            }
+
+            var castPath2 = Path.Combine(testDir, "integration-test-phase2.cast");
+            await _fixture.Container.CopyFromContainerAsync(recordPath2, castPath2);
+            if (File.Exists(castPath2))
+            {
+                _output.WriteLine($"Asciinema recording (phase 2): {castPath2}");
+            }
+        }
+        finally
+        {
+            // Try to stop aspire before container teardown
+            try
+            {
+                await _fixture.Container.ExecAsync(
+                    "export PATH=$HOME/.aspire/bin:$PATH && aspire stop 2>/dev/null || true",
+                    timeout: TimeSpan.FromSeconds(15));
+            }
+            catch
+            {
+                // Best effort cleanup
+            }
+
+            await _fixture.SaveTraceAsync(testDir, nameof(ExtensionShowsResourcesFromRunningAppHost));
+        }
+    }
+
+    private static string FindRepoRoot()
+    {
+        var dir = new DirectoryInfo(AppContext.BaseDirectory);
+        while (dir is not null)
+        {
+            if (File.Exists(Path.Combine(dir.FullName, "Aspire.slnx")))
+            {
+                return dir.FullName;
+            }
+            dir = dir.Parent;
+        }
+        return AppContext.BaseDirectory;
+    }
+
+    /// <summary>
+    /// Maximizes the VS Code terminal panel so it fills the entire editor area.
+    /// </summary>
+    private async Task MaximizeTerminalPanelAsync(IPage page)
+    {
+        _output.WriteLine("Maximizing terminal panel...");
+        await page.Keyboard.PressAsync("Control+Shift+KeyP");
+        await Task.Delay(500);
+        await page.Keyboard.TypeAsync("View: Toggle Maximized Panel", new() { Delay = 30 });
+        await Task.Delay(500);
+        await page.Keyboard.PressAsync("Enter");
+        await Task.Delay(1000);
+        _output.WriteLine("Terminal panel maximized");
+    }
+
+    private void DumpScreen(RemoteTerminalSession session, string label)
+    {
+        var lines = session.GetNonEmptyLines();
+        _output.WriteLine($"--- {label} ({lines.Count} non-empty lines) ---");
+        foreach (var line in lines)
+        {
+            _output.WriteLine($"  | {line}");
+        }
+        _output.WriteLine("---");
+    }
+
+    private static async Task ScreenshotAsync(IPage page, string testDir, string filename)
+    {
+        var path = Path.Combine(testDir, filename);
+        await page.ScreenshotAsync(new() { Path = path, FullPage = true });
+    }
+
+    public async ValueTask DisposeAsync()
+    {
+        await ValueTask.CompletedTask;
+    }
+}
